# addressregistry/src/main/java/com/teamdfx/addressregistry/dto/AddressDTO.java
package com.teamdfx.addressregistry.dto;

import lombok.Getter;
import lombok.Setter;

@Getter
@Setter
public class AddressDTO {
    private Long id;
    private String street;
    private String city;
    private String state;
    private String postalCode;
    private String country;
    private String addressType;

}